+++ lib/src/native/realm_core.dart
@@ -840,9 +840,6 @@ class _RealmCore {
       if (configuration.requestTimeout > 0) {
         _realmLib.realm_sync_client_config_set_connect_timeout(handle._pointer, configuration.requestTimeout);
       }
-
-      //_realmLib.realm_sync_config_set_resync_mode(handle._pointer, configuration.clientResyncMode.index);
-
       if (configuration.metadataEncryptionKey != null && configuration.metadataPersistenceMode == MetadataPersistenceMode.encrypted) {
         _realmLib.realm_sync_client_config_set_metadata_encryption_key(handle._pointer, configuration.metadataEncryptionKey!.toUint8Ptr(arena));
       }